requirements.txt
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
PyPDF2==3.0.1
pytesseract==0.3.13
